Clean up campaigns resolver tests a bit more
